Show all the code ignored by .gitignore

+++ src/main/java/com/cursojava/curso/controllers/UsuarioController.java
@@ -3,17 +3,19 @@ package com.cursojava.curso.controllers;
 
 import com.cursojava.curso.dao.UsuarioDao;
 import com.cursojava.curso.models.Usuario;
+import com.cursojava.curso.utils.JWTUtil;
 import de.mkammerer.argon2.Argon2;
 import de.mkammerer.argon2.Argon2Factory;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.web.bind.annotation.*;
 
-import java.util.ArrayList;
 import java.util.List;
 
 @RestController
 public class UsuarioController {
 
+    @Autowired
+    JWTUtil jwtUtil;
     @Autowired
     private UsuarioDao usuarioDao;
     @RequestMapping(value = "api/usuario/{id}")
@@ -23,8 +25,16 @@ public class UsuarioController {
     }
 
     @GetMapping(value = "api/usuarios")
-    public List<Usuario> getUsuarios() {
+    public List<Usuario> getUsuarios(@RequestHeader(value="Authorization") String token) {
+        if(!this.validarTokenJwt(token)) {
+            return null;
+        }
         return usuarioDao.getUsuarios();
+    }
+
+    public boolean validarTokenJwt(String token) {
+        String idUsuario = jwtUtil.getKey(token);
+        return idUsuario != null;
     }
     public Usuario editUsuario() {
         Usuario usuario = new Usuario();
@@ -37,7 +47,8 @@ public class UsuarioController {
     }
 
     @RequestMapping(value = "api/usuarios/{id}", method = RequestMethod.DELETE)
-    public void deleteUsuario(@PathVariable int id) {
+    public void  deleteUsuario(@PathVariable int id, @RequestHeader(value="Authorization") String token) {
+        if(!this.validarTokenJwt(token)) { return;}
         usuarioDao.deleteUsuario(id);
     }
